requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.3.0
django-filter==24.2
drf-spectacular==0.27.2
# For PostgreSQL (optional, SQLite works out of the box):
# psycopg2-binary==2.9.9
